+++ sdk/communication/azure-communication-identity/CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## 1.0.0-beta.6 (2021-03-09)
 ### Added
-- Added a retryPolicy() chain method to the `CommunicationIdentityClientBuilder`
+- Added a retryPolicy() chain method to the `CommunicationIdentityClientBuilder`.
+
+### Breaking
+- `CommunicationIdentityClient.createUserWithToken` and `CommunicationIdentityAsyncClient.createUserWithToken` have been renamed to
+`CommunicationIdentityClient.createUserAndToken` and `CommunicationIdentityAsyncClient.createUserAndToken`.
+- `CommunicationIdentityClient.createUserWithTokenWithResponse` and `CommunicationIdentityAsyncClient.createUserWithTokenWithResponse` have been renamed to
+`CommunicationIdentityClient.createUserAndTokenWithResponse` and `CommunicationIdentityAsyncClient.createUserAndTokenWithResponse`.
+- `CommunicationUserIdentifierWithTokenResult` class has been renamed to `CommunicationUserIdentifierAndTokenResult`.
 
 ## 1.0.0-beta.5 (2021-03-02)
 ### Breaking